requirements.txt
# Face Recognition System – Python Dependencies
# Install with:  pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3       # MTCNN + InceptionResnetV1 (FaceNet)
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
